+++ server/serverAPI/controller/UserController.ts
@@ -22,8 +22,8 @@ export default class UserController {
     }
 
     /**
-     * This property is a handler that is used for "getUsers" action of the user.
-     * It provides user with an ability to receive summary of non-sensitive information about all existing users in the database.
+     * Lets client to get information about all users existed on the server.
+     * Upon successful operation, this handler will return all users (including their non-sensitive information) existed on the server. 
      * 
      * @param req Request parameter that holds information about request
      * @param res Response parameter that holds information about response
@@ -34,9 +34,10 @@ export default class UserController {
     }
 
     /**
-     * This property is a handler that is used for "getUser" action of the user.
-     * It provides user with an ability to receive their own information (complete information) from the database.
-     * 
+     * Lets client to get information about user at specified userID.
+     * Upon successful operation, this handler will return complete information about specific user only if uid of the user with accessToken and uid of the 
+     * user at userID are the same.
+     *  
      * @param req Request parameter that holds information about request
      * @param res Response parameter that holds information about response
      */
@@ -63,12 +64,18 @@ export default class UserController {
             return;
         }
 
+        // if (user.uid !== req.params.uid) {
+        //     res.status(400).json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "Cannot access other people information."));
+        //     return;
+        // }
+
         res.status(200).json(ResponseFormatter.formatAsJSON(ResponseTypes.SUCCESS, user));
     }
 
     /**
-     * This property is a handler that is used for "updateUser" action of the user.
-     * It provides user with an ability to update their own information on the database.
+     * Lets client to update information of the user at specified userID.
+     * Upon successful operation, this handler will return updated user object only if uid of the user with accessToken and uid of the 
+     * user at userID are the same.
      * 
      * @param req Request parameter that holds information about request
      * @param res Response parameter that holds information about response
@@ -104,6 +111,11 @@ export default class UserController {
             return;
         }
 
+        // if (user.uid !== req.params.uid) {
+        //     res.status(400).json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "Cannot access other people information."));
+        //     return;
+        // }
+
         newUser.inventory = user.inventory;
 
         let updatedUser = await this.database.UpdateUser(userID, newUser);
@@ -112,8 +124,9 @@ export default class UserController {
     }
 
     /**
-     * This property is a handler that is used for "deleteUser" action of the user.
-     * It provides user with an ability to delete their own information from the database.
+     * Lets client to delete user object at specified userID.
+     * Upon successful operation, this handler will delete user object only if uid of the user with accessToken and uid of the 
+     * user at userID are the same.
      * 
      * @param req Request parameter that holds information about request
      * @param res Response parameter that holds information about response
@@ -129,6 +142,19 @@ export default class UserController {
             res.status(400).json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, logs));
             return;
         }
+
+        let user = await this.database.GetUser(new Map([["_id", userID]]));
+
+        if (user === null) {
+            res.status(404).json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "User hasn't been found."));
+            return;
+        }
+
+        // if (user.uid !== req.params.uid) {
+        //     res.status(400).json(ResponseFormatter.formatAsJSON(ResponseTypes.ERROR, "Cannot access other people information."));
+        //     return;
+        // }
+
         let result = await this.database.DeleteUser(userID);
 
         if (!result) {
@@ -140,8 +166,8 @@ export default class UserController {
     }
 
     /**
-     * This property is a handler that is used for "getFoods" action of the user.
-     * It provides user with an ability to retrieve information about their food inventory.
+     * Lets client to get all foods in user's inventory where user is at specified userID.
+     * Upon successful operation, this handler will return all food items in user's inventory.
      * 
      * @param req Request parameter that holds information about request
      * @param res Response parameter that holds information about response
@@ -173,8 +199,8 @@ export default class UserController {
     }
 
     /**
-    * This property is a handler that is used for "addFood" action of the user.
-    * It provides user with an ability to add food to the inventory.
+    * Lets client to add food to user's inventory where user is at specified userID.
+    * Upon successful operation, this handler will return all food items from user's inventory.
     * 
     * @param req Request parameter that holds information about request
     * @param res Response parameter that holds information about response
@@ -253,8 +279,8 @@ export default class UserController {
     }
 
     /**
-    * This property is a handler that is used for "getFood" action of the user.
-    * It provides user with an ability to retrieve information about specific food in their inventory.
+     * Lets client to get complete informations of the food item from user's inventory where user is at specified userID.
+     * Upon successful operation, this handler will return food item from user's inventory.
     * 
     * @param req Request parameter that holds information about request
     * @param res Response parameter that holds information about response
@@ -294,8 +320,8 @@ export default class UserController {
     }
 
     /**
-    * This property is a handler that is used for "updateFood" action of the user.
-    * It provides user with an ability to update food in the user's inventory.
+     * Lets client to update information of the food item from user's inventory where user is at specified userID.
+     * Upon successful operation, this handler will return all food items in user's inventory.
     * 
     * @param req Request parameter that holds information about request
     * @param res Response parameter that holds information about response
@@ -389,8 +415,8 @@ export default class UserController {
     }
 
     /**
-    * This property is a handler that is used for "deleteFood" action of the user.
-    * It provides user with an ability to delete specific food from their inventory.
+    * Lets client to delete food item from user's inventory where user is at specified userID.
+    * Upon successful operation, this handler will return all food items in user's inventory.
     * 
     * @param req Request parameter that holds information about request
     * @param res Response parameter that holds information about response
